+++ lib/segment/src/lib.rs
@@ -5,6 +5,7 @@ pub mod fixtures;
 pub mod id_tracker;
 pub mod index;
 pub mod payload_storage;
+pub mod pending_changes;
 pub mod segment;
 pub mod segment_constructor;
 pub mod spaces;
